Update composer to v2, move classes to support PSR-4 autoloading
