+++ .phpstorm.meta.php
@@ -171,7 +171,7 @@ namespace PHPSTORM_META {
             'migration.repository' => \Illuminate\Database\Migrations\DatabaseMigrationRepository::class,
             'migrator' => \Illuminate\Database\Migrations\Migrator::class,
             'queue' => \Illuminate\Queue\QueueManager::class,
-            'queue.connection' => \Illuminate\Queue\SyncQueue::class,
+            'queue.connection' => \Laravel\Horizon\RedisQueue::class,
             'queue.failer' => \Illuminate\Queue\Failed\DatabaseFailedJobProvider::class,
             'queue.listener' => \Illuminate\Queue\Listener::class,
             'queue.worker' => \Illuminate\Queue\Worker::class,
@@ -356,7 +356,7 @@ namespace PHPSTORM_META {
             'migration.repository' => \Illuminate\Database\Migrations\DatabaseMigrationRepository::class,
             'migrator' => \Illuminate\Database\Migrations\Migrator::class,
             'queue' => \Illuminate\Queue\QueueManager::class,
-            'queue.connection' => \Illuminate\Queue\SyncQueue::class,
+            'queue.connection' => \Laravel\Horizon\RedisQueue::class,
             'queue.failer' => \Illuminate\Queue\Failed\DatabaseFailedJobProvider::class,
             'queue.listener' => \Illuminate\Queue\Listener::class,
             'queue.worker' => \Illuminate\Queue\Worker::class,
@@ -541,7 +541,7 @@ namespace PHPSTORM_META {
             'migration.repository' => \Illuminate\Database\Migrations\DatabaseMigrationRepository::class,
             'migrator' => \Illuminate\Database\Migrations\Migrator::class,
             'queue' => \Illuminate\Queue\QueueManager::class,
-            'queue.connection' => \Illuminate\Queue\SyncQueue::class,
+            'queue.connection' => \Laravel\Horizon\RedisQueue::class,
             'queue.failer' => \Illuminate\Queue\Failed\DatabaseFailedJobProvider::class,
             'queue.listener' => \Illuminate\Queue\Listener::class,
             'queue.worker' => \Illuminate\Queue\Worker::class,
@@ -726,7 +726,7 @@ namespace PHPSTORM_META {
             'migration.repository' => \Illuminate\Database\Migrations\DatabaseMigrationRepository::class,
             'migrator' => \Illuminate\Database\Migrations\Migrator::class,
             'queue' => \Illuminate\Queue\QueueManager::class,
-            'queue.connection' => \Illuminate\Queue\SyncQueue::class,
+            'queue.connection' => \Laravel\Horizon\RedisQueue::class,
             'queue.failer' => \Illuminate\Queue\Failed\DatabaseFailedJobProvider::class,
             'queue.listener' => \Illuminate\Queue\Listener::class,
             'queue.worker' => \Illuminate\Queue\Worker::class,
@@ -911,7 +911,7 @@ namespace PHPSTORM_META {
             'migration.repository' => \Illuminate\Database\Migrations\DatabaseMigrationRepository::class,
             'migrator' => \Illuminate\Database\Migrations\Migrator::class,
             'queue' => \Illuminate\Queue\QueueManager::class,
-            'queue.connection' => \Illuminate\Queue\SyncQueue::class,
+            'queue.connection' => \Laravel\Horizon\RedisQueue::class,
             'queue.failer' => \Illuminate\Queue\Failed\DatabaseFailedJobProvider::class,
             'queue.listener' => \Illuminate\Queue\Listener::class,
             'queue.worker' => \Illuminate\Queue\Worker::class,
@@ -1096,7 +1096,7 @@ namespace PHPSTORM_META {
             'migration.repository' => \Illuminate\Database\Migrations\DatabaseMigrationRepository::class,
             'migrator' => \Illuminate\Database\Migrations\Migrator::class,
             'queue' => \Illuminate\Queue\QueueManager::class,
-            'queue.connection' => \Illuminate\Queue\SyncQueue::class,
+            'queue.connection' => \Laravel\Horizon\RedisQueue::class,
             'queue.failer' => \Illuminate\Queue\Failed\DatabaseFailedJobProvider::class,
             'queue.listener' => \Illuminate\Queue\Listener::class,
             'queue.worker' => \Illuminate\Queue\Worker::class,
@@ -1281,7 +1281,7 @@ namespace PHPSTORM_META {
             'migration.repository' => \Illuminate\Database\Migrations\DatabaseMigrationRepository::class,
             'migrator' => \Illuminate\Database\Migrations\Migrator::class,
             'queue' => \Illuminate\Queue\QueueManager::class,
-            'queue.connection' => \Illuminate\Queue\SyncQueue::class,
+            'queue.connection' => \Laravel\Horizon\RedisQueue::class,
             'queue.failer' => \Illuminate\Queue\Failed\DatabaseFailedJobProvider::class,
             'queue.listener' => \Illuminate\Queue\Listener::class,
             'queue.worker' => \Illuminate\Queue\Worker::class,
@@ -1466,7 +1466,7 @@ namespace PHPSTORM_META {
             'migration.repository' => \Illuminate\Database\Migrations\DatabaseMigrationRepository::class,
             'migrator' => \Illuminate\Database\Migrations\Migrator::class,
             'queue' => \Illuminate\Queue\QueueManager::class,
-            'queue.connection' => \Illuminate\Queue\SyncQueue::class,
+            'queue.connection' => \Laravel\Horizon\RedisQueue::class,
             'queue.failer' => \Illuminate\Queue\Failed\DatabaseFailedJobProvider::class,
             'queue.listener' => \Illuminate\Queue\Listener::class,
             'queue.worker' => \Illuminate\Queue\Worker::class,
@@ -1651,7 +1651,7 @@ namespace PHPSTORM_META {
             'migration.repository' => \Illuminate\Database\Migrations\DatabaseMigrationRepository::class,
             'migrator' => \Illuminate\Database\Migrations\Migrator::class,
             'queue' => \Illuminate\Queue\QueueManager::class,
-            'queue.connection' => \Illuminate\Queue\SyncQueue::class,
+            'queue.connection' => \Laravel\Horizon\RedisQueue::class,
             'queue.failer' => \Illuminate\Queue\Failed\DatabaseFailedJobProvider::class,
             'queue.listener' => \Illuminate\Queue\Listener::class,
             'queue.worker' => \Illuminate\Queue\Worker::class,
@@ -1836,7 +1836,7 @@ namespace PHPSTORM_META {
             'migration.repository' => \Illuminate\Database\Migrations\DatabaseMigrationRepository::class,
             'migrator' => \Illuminate\Database\Migrations\Migrator::class,
             'queue' => \Illuminate\Queue\QueueManager::class,
-            'queue.connection' => \Illuminate\Queue\SyncQueue::class,
+            'queue.connection' => \Laravel\Horizon\RedisQueue::class,
             'queue.failer' => \Illuminate\Queue\Failed\DatabaseFailedJobProvider::class,
             'queue.listener' => \Illuminate\Queue\Listener::class,
             'queue.worker' => \Illuminate\Queue\Worker::class,
@@ -2021,7 +2021,7 @@ namespace PHPSTORM_META {
             'migration.repository' => \Illuminate\Database\Migrations\DatabaseMigrationRepository::class,
             'migrator' => \Illuminate\Database\Migrations\Migrator::class,
             'queue' => \Illuminate\Queue\QueueManager::class,
-            'queue.connection' => \Illuminate\Queue\SyncQueue::class,
+            'queue.connection' => \Laravel\Horizon\RedisQueue::class,
             'queue.failer' => \Illuminate\Queue\Failed\DatabaseFailedJobProvider::class,
             'queue.listener' => \Illuminate\Queue\Listener::class,
             'queue.worker' => \Illuminate\Queue\Worker::class,
